Experiscanner tsx (#90278)
## About The Pull Request
Just made it a typescript ui in the process of fixing the scroll wheel

Fixes #90308 (dupe)
Fixes #90272 (dupe)
Fixes #90265 

Removed flex (deprecated)
## Why It's Good For The Game
it scrolle

![T8N1N18SZP](https://github.com/user-attachments/assets/016d4b3b-7d14-449a-87da-53fce1d1e178)

## Changelog
:cl:
fix: Fixed scrolling the experiscanner UI
/:cl:
